+++ warpdemux/logger.py
@@ -1,0 +1,34 @@
+"""
+WarpDemuX
+
+Copyright (c) 2023 by Wiep K. van der Toorn
+Contact: w.vandertoorn@fu-berlin.de
+
+"""
+
+import sys
+import logging
+
+
+# Set up logging to both file and stdout
+class DualLogger:
+    def __init__(self, filename):
+        self.terminal = sys.stdout
+        self.log = open(filename, "a")
+
+    def write(self, message):
+        self.terminal.write(message)
+        self.log.write(message)
+
+    def flush(self):
+        self.terminal.flush()
+        self.log.flush()
+
+
+def setup_logger(filename="warpdemux.log"):
+    logging.basicConfig(
+        level=logging.INFO,
+        format="%(asctime)s - %(levelname)s - %(message)s",
+        datefmt="%Y-%m-%d %H:%M:%S",
+        stream=DualLogger(filename),
+    )